Add MatchPlayer back.
Added MatchPlayer back to have a *<-->* relationship between
Player and Teams. The player can be a part of multiple teams
and team can have multiple players. The problem is explained
in my blog post
https://www.naresh.se/2015/07/01/ormlite-lightweight-orm-package-android-relationship/
The final solution for the time being is to add an
intermediary table. A default edit screen is also added.